+++ how-to/workspace-platform-starter/client/src/framework/shapes/favorite-shapes.ts
@@ -1,20 +1,51 @@
 import type { PlatformStorageMetadata } from "./platform-shapes";
 
 /**
+ * Favorite type for App.
+ */
+export const FAVORITE_TYPE_NAME_APP = "app";
+
+/**
+ * Favorite type for Workspace.
+ */
+export const FAVORITE_TYPE_NAME_WORKSPACE = "workspace";
+
+/**
+ * Favorite type for Page.
+ */
+export const FAVORITE_TYPE_NAME_PAGE = "page";
+
+/**
+ * Favorite type for Query.
+ */
+export const FAVORITE_TYPE_NAME_QUERY = "query";
+
+/**
  * Names for all the favorite types.
  */
-export type FavoriteTypeNames = "app" | "workspace" | "page" | "query";
+export type FavoriteTypeNames =
+	| typeof FAVORITE_TYPE_NAME_APP
+	| typeof FAVORITE_TYPE_NAME_WORKSPACE
+	| typeof FAVORITE_TYPE_NAME_PAGE
+	| typeof FAVORITE_TYPE_NAME_QUERY;
 
 /**
  * Options for the favorite provider.
  */
 export interface FavoriteProviderOptions {
 	/**
+	 * Is the provider enabled, defaults to true.
+	 */
+	enabled?: boolean;
+
+	/**
 	 * The icon that should be used if you want to indicate this is a favorite action
 	 */
 	favoriteIcon: string;
 
-	/** The icon to use to indicate that this favorite can be unset */
+	/**
+	 * The icon to use to indicate that this favorite can be unset
+	 */
 	unfavoriteIcon: string;
 
 	/**
@@ -51,7 +82,7 @@ export interface FavoriteEntry {
 	/**
 	 * The timestamp for the entry.
 	 */
-	timestamp: Date;
+	timestamp?: Date;
 
 	/**
 	 * Does this favorite have a suggested label that can be used to avoid a lookup
@@ -104,7 +135,7 @@ export interface FavoriteClient {
 	 * @param byType the type of saved favorite you are looking for
 	 * @returns An array of saved favorites or an empty array if it was unable to get any back
 	 */
-	getSavedFavorites(byType: FavoriteTypeNames): Promise<FavoriteEntry[] | undefined>;
+	getSavedFavorites(byType?: FavoriteTypeNames): Promise<FavoriteEntry[] | undefined>;
 	/**
 	 * The ability to request a particular saved favorite.
 	 * @param id the id of the favorite you are looking for